fix: prebuild design-tokens package
