+++ libraries/design-tokens/package.json
@@ -3,17 +3,22 @@
   "version": "0.0.0",
   "license": "UNLICENSED",
   "sideEffects": false,
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
   "exports": {
     ".": {
-      "import": "./src/index.ts"
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
     },
     "./themes/*": {
-      "import": "./themes/*.ts"
+      "import": "./dist/themes/*.js",
+      "types": "./dist/themes/*.d.ts"
     }
   },
+  "files": [
+    "dist"
+  ],
   "scripts": {
+    "build": "tsc --build && cp -r src/assets dist",
+    "dev": "tsc --build --watch --preserveWatchOutput",
     "codegen": "tsx scripts/codegen"
   },
   "devDependencies": {
