+++ libraries/design-tokens/scripts/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "@yas/tsconfig/node.json",
+  "compilerOptions": {
+    "rootDir": "."
+  }
+}
