+++ libraries/design-tokens/src/assets/inter/ttf/inter-latin-100-normal.ttf
