+++ libraries/design-tokens/src/assets/inter/ttf/inter-latin-200-normal.ttf
