+++ libraries/design-tokens/src/assets/inter/ttf/inter-latin-300-normal.ttf
