+++ libraries/design-tokens/src/assets/inter/ttf/inter-latin-400-normal.ttf
