+++ libraries/design-tokens/src/assets/inter/ttf/inter-latin-500-normal.ttf
