+++ libraries/design-tokens/src/assets/inter/ttf/inter-latin-600-normal.ttf
