+++ libraries/design-tokens/src/assets/inter/ttf/inter-latin-700-normal.ttf
