+++ libraries/design-tokens/src/assets/inter/ttf/inter-latin-800-normal.ttf
