+++ libraries/design-tokens/src/assets/inter/ttf/inter-latin-900-normal.ttf
