+++ libraries/design-tokens/src/assets/inter/webfonts/inter-latin-100-normal.woff2
