+++ libraries/design-tokens/src/assets/inter/webfonts/inter-latin-200-normal.woff2
