+++ libraries/design-tokens/src/assets/inter/webfonts/inter-latin-300-normal.woff2
