+++ libraries/design-tokens/src/assets/inter/webfonts/inter-latin-400-normal.woff2
