+++ libraries/design-tokens/src/assets/inter/webfonts/inter-latin-500-normal.woff2
