+++ libraries/design-tokens/src/assets/inter/webfonts/inter-latin-600-normal.woff
