+++ libraries/design-tokens/src/assets/inter/webfonts/inter-latin-700-normal.woff2
