+++ libraries/design-tokens/src/assets/inter/webfonts/inter-latin-800-normal.woff2
