+++ libraries/design-tokens/src/custom.ts
@@ -1,6 +1,6 @@
 import { inter } from "./fonts/inter";
 import type * as shared from "./shared";
-import type { FontFaceDefinition } from "./fonts/convention";
+import type { FontFaceDefinition } from "./internal";
 
 /**
  * This module implements tokens that do not exist in figma or that cannot be automatically generated.
